// src/generated/protobuf/v1/lobby.ts
export namespace PBTypes {
    export enum LobbyStatus {
        LOBBY_OPEN = 0,
        LOBBY_FULL = 1,
        LOBBY_STARTING = 2,
        LOBBY_STARTED = 3,
        LOBBY_CLOSED = 4
    }
// ...
    export class LobbyList extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            lobbies?: Lobby[];
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [1], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("lobbies" in data && data.lobbies != undefined) {
                    this.lobbies = data.lobbies;
                }
            }
        }
        get lobbies() {
            return pb_1.Message.getRepeatedWrapperField(this, Lobby, 1) as Lobby[];
        }
        set lobbies(value: Lobby[]) {
            pb_1.Message.setRepeatedWrapperField(this, 1, value);
        }
        static fromObject(data: {
            lobbies?: ReturnType<typeof Lobby.prototype.toObject>[];
        }): LobbyList {
            const message = new LobbyList({});
            if (data.lobbies != null) {
                message.lobbies = data.lobbies.map(item => Lobby.fromObject(item));
            }
            return message;
        }
        toObject() {
            const data: {
                lobbies?: ReturnType<typeof Lobby.prototype.toObject>[];
            } = {};
            if (this.lobbies != null) {
                data.lobbies = this.lobbies.map((item: Lobby) => item.toObject());
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.lobbies.length)
                writer.writeRepeatedMessage(1, this.lobbies, (item: Lobby) => item.serialize(writer));
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): LobbyList {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new LobbyList();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        reader.readMessage(message.lobbies, () => pb_1.Message.addToRepeatedWrapperField(message, 1, Lobby.deserialize(reader), Lobby));
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): LobbyList {
            return LobbyList.deserialize(bytes);
        }
    }
    export class CreateLobbyRequest extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            name?: string;
            is_private?: boolean;
            pin?: string;
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("name" in data && data.name != undefined) {
                    this.name = data.name;
                }
                if ("is_private" in data && data.is_private != undefined) {
                    this.is_private = data.is_private;
                }
                if ("pin" in data && data.pin != undefined) {
                    this.pin = data.pin;
                }
            }
        }
        get name() {
            return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
        }
        set name(value: string) {
            pb_1.Message.setField(this, 1, value);
        }
        get is_private() {
            return pb_1.Message.getFieldWithDefault(this, 2, false) as boolean;
        }
        set is_private(value: boolean) {
            pb_1.Message.setField(this, 2, value);
        }
        get pin() {
            return pb_1.Message.getFieldWithDefault(this, 3, "") as string;
        }
        set pin(value: string) {
            pb_1.Message.setField(this, 3, value);
        }
        static fromObject(data: {
            name?: string;
            is_private?: boolean;
            pin?: string;
        }): CreateLobbyRequest {
            const message = new CreateLobbyRequest({});
            if (data.name != null) {
                message.name = data.name;
            }
            if (data.is_private != null) {
                message.is_private = data.is_private;
            }
            if (data.pin != null) {
                message.pin = data.pin;
            }
            return message;
        }
        toObject() {
            const data: {
                name?: string;
                is_private?: boolean;
                pin?: string;
            } = {};
            if (this.name != null) {
                data.name = this.name;
            }
            if (this.is_private != null) {
                data.is_private = this.is_private;
            }
            if (this.pin != null) {
                data.pin = this.pin;
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.name.length)
                writer.writeString(1, this.name);
            if (this.is_private != false)
                writer.writeBool(2, this.is_private);
            if (this.pin.length)
                writer.writeString(3, this.pin);
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): CreateLobbyRequest {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new CreateLobbyRequest();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        message.name = reader.readString();
                        break;
                    case 2:
                        message.is_private = reader.readBool();
                        break;
                    case 3:
                        message.pin = reader.readString();
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): CreateLobbyRequest {
            return CreateLobbyRequest.deserialize(bytes);
        }
    }
    export class JoinLobbyRequest extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            pin?: string;
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("pin" in data && data.pin != undefined) {
                    this.pin = data.pin;
                }
            }
        }
        get pin() {
            return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
        }
        set pin(value: string) {
            pb_1.Message.setField(this, 1, value);
        }
        static fromObject(data: {
            pin?: string;
        }): JoinLobbyRequest {
            const message = new JoinLobbyRequest({});
            if (data.pin != null) {
                message.pin = data.pin;
            }
            return message;
        }
        toObject() {
            const data: {
                pin?: string;
            } = {};
            if (this.pin != null) {
                data.pin = this.pin;
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.pin.length)
                writer.writeString(1, this.pin);
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): JoinLobbyRequest {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new JoinLobbyRequest();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        message.pin = reader.readString();
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): JoinLobbyRequest {
            return JoinLobbyRequest.deserialize(bytes);
        }
    }
    export class ReadyRequest extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            ready?: boolean;
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("ready" in data && data.ready != undefined) {
                    this.ready = data.ready;
                }
            }
        }
        get ready() {
            return pb_1.Message.getFieldWithDefault(this, 1, false) as boolean;
        }
        set ready(value: boolean) {
            pb_1.Message.setField(this, 1, value);
        }
        static fromObject(data: {
            ready?: boolean;
        }): ReadyRequest {
            const message = new ReadyRequest({});
            if (data.ready != null) {
                message.ready = data.ready;
            }
            return message;
        }
        toObject() {
            const data: {
                ready?: boolean;
            } = {};
            if (this.ready != null) {
                data.ready = this.ready;
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.ready != false)
                writer.writeBool(1, this.ready);
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): ReadyRequest {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new ReadyRequest();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        message.ready = reader.readBool();
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): ReadyRequest {
            return ReadyRequest.deserialize(bytes);
        }
    }
}